+++ packages/mint-components/src/components/sqm-rewards-table/RewardsTableCell.stories.tsx
@@ -432,23 +432,15 @@ export const StatusCellExpired = () => {
 const pending = {
   statuses: ["PENDING"],
 };
-// const pendingTaxReview = {
-//   statuses: ["PENDING_TAX_REVIEW"],
-// };
-// const pendingNewTaxForm = {
-//   statuses: ["PENDING_NEW_TAX_FORM"],
-// };
-// const pendingTaxSubmission = {
-//   statuses: ["PENDING_TAX_SUBMISSION"],
-// };
-// const pendingPartnerCreation = {
-//   statuses: ["PENDING_PARTNER_CREATION"],
-// };
 const payoutSent = {
   statuses: ["PAYOUT_SENT"],
 };
 const payoutFailed = {
   statuses: ["PAYOUT_FAILED"],
+};
+const payoutCancelled = {
+  statuses: ["PAYOUT_CANCELLED"],
+  dateCancelled: 1355612521321,
 };
 const us_tax = {
   pendingReasons: ["US_TAX"],
@@ -554,6 +546,15 @@ export const StatusCellPayoutFailed = () => {
     <sqm-rewards-table-status-cell
       statusText="Payout Failed"
       reward={{ ...rewardsData, ...payoutFailed }}
+    ></sqm-rewards-table-status-cell>
+  );
+};
+
+export const StatusCellPayoutCancelled = () => {
+  return (
+    <sqm-rewards-table-status-cell
+      statusText="Payout Cancelled"
+      reward={{ ...rewardsData, ...payoutCancelled }}
     ></sqm-rewards-table-status-cell>
   );
 };