+++ packages/mint-components/src/components/sqm-widget-verification/useWidgetVerification.ts
@@ -1,0 +1,70 @@
+import {
+  useLazyQuery,
+  useParentState,
+  useSetParent,
+  useUserIdentity,
+} from "@saasquatch/component-boilerplate";
+import { useState } from "@saasquatch/stencil-hooks";
+import { useEffect } from "@saasquatch/universal-hooks";
+import { gql } from "graphql-request";
+import {
+  SHOW_CODE_NAMESPACE,
+  VERIFICATION_EMAIL_NAMESPACE,
+  VERIFICATION_PARENT_NAMESPACE,
+} from "./keys";
+
+const USER_LOOKUP = gql`
+  query checkUserVerification {
+    viewer {
+      ... on User {
+        id
+        accountId
+        email
+        emailVerified
+        managedIdentity {
+          email
+          emailVerified
+        }
+      }
+    }
+  }
+`;
+
+export function useWidgetVerification() {
+  const userIdentity = useUserIdentity();
+  const [showCode, setShowCode] = useParentState<boolean>({
+    namespace: SHOW_CODE_NAMESPACE,
+    initialValue: false,
+  });
+  const [email, setEmail] = useParentState<string | undefined>({
+    namespace: VERIFICATION_EMAIL_NAMESPACE,
+    initialValue: userIdentity?.email,
+  });
+  const setContext = useSetParent(VERIFICATION_PARENT_NAMESPACE);
+  const [loading, setLoading] = useState(true);
+  const [fetch] = useLazyQuery(USER_LOOKUP);
+
+  useEffect(() => {
+    const checkUser = async () => {
+      try {
+        const res = await fetch({});
+        if (!res || res instanceof Error) throw new Error();
+
+        if (res?.viewer?.emailVerified) setContext(true);
+        else if (res?.viewer?.managedIdentity?.emailVerified) setContext(true);
+      } catch (e) {
+        console.error("Could not fetch user information:", e);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    checkUser();
+  }, []);
+
+  const onVerification = () => {
+    setContext(true);
+  };
+
+  return { showCode, onVerification, loading };
+}