+++ packages/mint-components/src/components/tax-and-cash/sqm-payout-details-card/PayoutDetailsCard.stories.tsx
@@ -16,18 +16,6 @@ export const Loading = () => {
   );
 };
 
-export const Error = () => {
-  return (
-    <sqm-payout-details-card
-      demoData={{
-        states: {
-          error: true,
-        },
-      }}
-    ></sqm-payout-details-card>
-  );
-};
-
 export const NextPayout = () => {
   return (
     <sqm-payout-details-card