+++ packages/mint-components/src/components/tax-and-cash/sqm-user-info-form/sqm-user-info-form-view.tsx
@@ -100,6 +100,7 @@ export interface UserInfoFormViewProps {
     termsAndConditionsLabel: string;
     termsAndConditionsLink: string;
     taxAndPayoutsDescription: string;
+    supportLink: string;
     error: {
       generalTitle: string;
       generalDescription: string;
@@ -204,10 +205,8 @@ const style = {
 
   PhoneInputsSection: {
     display: "flex",
-    alignItems: "flex-start",
     flexDirection: "column",
     gap: "4px",
-
     "& p": {
       fontSize: "var(--sl-font-size-small)",
       color: "var(--sl-input-label-color)",
@@ -217,7 +216,6 @@ const style = {
 
   PhoneInputsContainer: {
     display: "flex",
-    alignItems: "flex-start",
     gap: "4px",
     width: "100%",
 
@@ -319,7 +317,22 @@ export const UserInfoFormView = (props: UserInfoFormViewProps) => {
             <sl-icon slot="icon" name="exclamation-octagon"></sl-icon>
             <strong>{text.error.loadingErrorAlertHeader}</strong>
             <br />
-            {text.error.loadingErrorAlertDescription}
+            {intl.formatMessage(
+              {
+                id: "loadingErrorAlertDescription",
+                defaultMessage: text.error.loadingErrorAlertDescription,
+              },
+              {
+                supportLink: (
+                  <a
+                    target="_blank"
+                    href={`mailto:advocate-support@impact.com`}
+                  >
+                    {text.supportLink}
+                  </a>
+                ),
+              }
+            )}
           </sl-alert>
           <br />
         </div>
@@ -357,7 +370,22 @@ export const UserInfoFormView = (props: UserInfoFormViewProps) => {
               <sl-icon slot="icon" name="exclamation-octagon"></sl-icon>
               <strong>{text.error.generalTitle}</strong>
               <br />
-              {text.error.generalDescription}
+              {intl.formatMessage(
+                {
+                  id: "generalDescription",
+                  defaultMessage: text.error.generalDescription,
+                },
+                {
+                  supportLink: (
+                    <a
+                      target="_blank"
+                      href={`mailto:advocate-support@impact.com`}
+                    >
+                      {text.supportLink}
+                    </a>
+                  ),
+                }
+              )}
             </sl-alert>
           )}
           {(states.isPartner || states.isUser) && (
@@ -370,7 +398,22 @@ export const UserInfoFormView = (props: UserInfoFormViewProps) => {
               <sl-icon slot="icon" name="info-circle"></sl-icon>
               <strong>{text.isPartnerAlertHeader}</strong>
               <br />
-              {text.isPartnerAlertDescription}
+              {intl.formatMessage(
+                {
+                  id: "isPartnerAlertDescription",
+                  defaultMessage: text.isPartnerAlertDescription,
+                },
+                {
+                  supportLink: (
+                    <a
+                      target="_blank"
+                      href={`mailto:advocate-support@impact.com`}
+                    >
+                      {text.supportLink}
+                    </a>
+                  ),
+                }
+              )}
             </sl-alert>
           )}
 
@@ -525,14 +568,14 @@ export const UserInfoFormView = (props: UserInfoFormViewProps) => {
                     id="phoneNumber"
                     name="/phoneNumber"
                     value={formState.phoneNumber}
-                    validationError={({ value }) =>
+                    validationError={({ value }) => {
                       // Naive phone number validation
                       validateBillingField(/[a-zA-Z]+/, value) &&
-                      formatErrorMessage(
-                        text.phoneNumber,
-                        text.error.fieldInvalidError
-                      )
-                    }
+                        formatErrorMessage(
+                          text.phoneNumber,
+                          text.error.fieldInvalidError
+                        );
+                    }}
                     disabled={states.disabled || states.isPartner}
                     {...(formState.errors?.phoneNumber
                       ? {