+++ packages/component-environment/CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.9] - 2025-03-18
+## [1.0.10] - 2025-03-18
 
 ### Changed
 
 - Add `googleClientId` to `PortalEnv` type 
+
+## [1.0.9] - 2025-03-12
+
+### Changed
+
+- Added handling for `email` field to the JWT methods
 
 ## [1.0.8] - 2024-05-31