+++ packages/mint-components/CHANGELOG.md
@@ -16,6 +16,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - \<sqm-portal-google-login>
   - Adds the ability to login with a Google account 
 
+## [1.12.0] - 2025-03-12
+
+### Added
+- \<sqm-widget-verification>
+  - Adds a verification gate to a given slot, requiring a user passes a 2FA flow to view the component.
+
+### Changed
+- \<sqm-tax-and-cash>
+  - Added a 2FA step required to update withdrawal settings.
+
 ## [1.11.1] - 2025-02-18
 
 ### Added
@@ -1105,6 +1115,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [unreleased]: https://github.com/saasquatch/program-tools/compare/mint-components@1.13.0...HEAD
 [1.13.0]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.13.0
+[1.12.0]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.12.0
 [1.11.1]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.11.1
 [1.11.0]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.11.0
 [1.10.4]: https://github.com/saasquatch/program-tools/releases/tag/%40saasquatch%2Fmint-components%401.10.4